requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
asyncpg==0.27.0
backcall==0.2.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
decorator==5.1.1
executing==1.2.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
traitlets==5.9.0
typing_extensions==4.5.0
uvicorn==0.22.0
wcwidth==0.2.6
